Fix type promotion in math function

+++ 3d-viewer/3d_rendering/3d_render_raytracing/PerlinNoise.cpp
@@ -40,6 +40,7 @@
  */
 
 #include "PerlinNoise.h"
+#include <cmath>
 #include <iostream>
 #include <cmath>
 #include <random>
@@ -97,14 +98,14 @@ PerlinNoise::PerlinNoise( unsigned int seed )
 float PerlinNoise::noise( float x, float y, float z ) const
 {
     // Find the unit cube that contains the point
-    int X = (int) ((float)floor( x )) & 255;
-    int Y = (int) ((float)floor( y )) & 255;
-    int Z = (int) ((float)floor( z )) & 255;
+    int X = static_cast<int>( std::floor( x ) ) & 255;
+    int Y = static_cast<int>( std::floor( y ) ) & 255;
+    int Z = static_cast<int>( std::floor( z ) ) & 255;
 
     // Find relative x, y,z of point in cube
-    x -= (float)floor( x );
-    y -= (float)floor( y );
-    z -= (float)floor( z );
+    x -= std::floor( x );
+    y -= std::floor( y );
+    z -= std::floor( z );
 
     // Compute fade curves for each of x, y, z
     const float u = fade( x );
@@ -143,12 +144,12 @@ float PerlinNoise::noise( float x, float y, float z ) const
 float PerlinNoise::noise( float x, float y ) const
 {
     // Find the unit cube that contains the point
-    int X = (int) ((float)floor( x )) & 255;
-    int Y = (int) ((float)floor( y )) & 255;
+    int X = static_cast<int>( std::floor( x ) ) & 255;
+    int Y = static_cast<int>( std::floor( y ) ) & 255;
 
     // Find relative x, y,z of point in cube
-    x -= (float)floor( x );
-    y -= (float)floor( y );
+    x -= std::floor( x );
+    y -= std::floor( y );
 
     // Compute fade curves for each of x, y
     const float u = fade( x );
